+++ src/Sulu/Bundle/MediaBundle/DependencyInjection/SuluMediaExtension.php
@@ -43,9 +43,8 @@ class SuluMediaExtension extends Extension
         $container->setParameter('sulu_media.collection.type.default', array(
             'id' => 1
         ));
+        $container->setParameter('sulu_media.collection.previews.format', '50x50');
         $container->setParameter('sulu_media.format_manager.config_paths', $config['format_manager']['config_paths']);
-        $container->setParameter('sulu_media.collection.previews.limit', 3);
-        $container->setParameter('sulu_media.collection.previews.format', '150x100');
         $container->setParameter('sulu_media.media.max_file_size', '16MB');
         $container->setParameter('sulu_media.media.blocked_file_types', $config['format_manager']['blocked_file_types']);
         $container->setParameter('sulu_media.media.storage.local.path', '%kernel.root_dir%/../uploads/media');