+++ src/Sulu/Bundle/MediaBundle/Resources/public/js/components/collections/components/media-edit/main.js
@@ -67,6 +67,14 @@ define(function () {
             return createEventName.call(this, 'closed');
         },
 
+        /**
+         * raised when component is initialized
+         * @event sulu.media-edit.closed
+         */
+        INITIALIZED = function () {
+            return createEventName.call(this, 'initialized');
+        },
+
         /** returns normalized event names */
         createEventName = function (postFix) {
             return namespace + (this.options.instanceName ? this.options.instanceName + '.' : '') + postFix;
@@ -102,6 +110,8 @@ define(function () {
             // stores the multiple edit-form
             this.$multiple = null;
             this.startLoadingOverlay();
+
+            this.sandbox.emit(INITIALIZED.call(this));
         },
 
         /**
@@ -119,12 +129,6 @@ define(function () {
             this.sandbox.on(LOADING.call(this), function() {
                 this.sandbox.emit('husky.overlay.media-edit.loading.open');
             }.bind(this));
-
-            // change language (single-edit)
-            this.sandbox.on('husky.overlay.media-edit.language-changed', this.languageChangedSingle.bind(this));
-
-            // change language (multi-edit)
-            this.sandbox.on('husky.overlay.media-multiple-edit.language-changed', this.languageChangedMultiple.bind(this));
         },
 
         /**
@@ -137,56 +141,6 @@ define(function () {
             } else {
                 this.editSingleMedia(media);
             }
-        },
-
-        /**
-         * Handles the changing of the language in the single-edit overlay
-         * @param locale
-         */
-        languageChangedSingle: function (locale) {
-            this.sandbox.emit('sulu.media.collections.reload-single-media',
-                this.media.id, {locale: locale},
-                function (media) {
-                    this.media = media;
-                    this.sandbox.form.setData(constants.infoFormSelector, this.media);
-                }.bind(this)
-            );
-        },
-
-        /**
-         * Handles the changing of the language in the single-edit overlay
-         * @param locale
-         */
-        languageChangedMultiple: function(locale) {
-            this.sandbox.emit('sulu.media.collections.reload-media',
-                this.medias, {locale: locale},
-                function(medias) {
-                    this.medias = medias;
-                    var descriptionVisible = this.sandbox.dom.is(
-                        constants.multipleEditFormSelector + ' ' + constants.multipleEditDescSelector, ':visible'
-                    ),
-                        tagsVisible = this.sandbox.dom.is(
-                            constants.multipleEditFormSelector + ' ' + constants.multipleEditTagsSelector, ':visible'
-                        );
-
-                    this.sandbox.stop(constants.multipleEditFormSelector + ' *');
-                    this.sandbox.form.setData(constants.multipleEditFormSelector, {
-                        records: this.medias
-                    }).then(function() {
-                            this.sandbox.start(constants.multipleEditFormSelector);
-                            if (descriptionVisible === true) {
-                                this.sandbox.dom.show(
-                                    constants.multipleEditFormSelector + ' ' + constants.multipleEditDescSelector
-                                );
-                            }
-                            if (tagsVisible) {
-                                this.sandbox.dom.show(
-                                    constants.multipleEditFormSelector + ' ' + constants.multipleEditTagsSelector
-                                );
-                            }
-                        }.bind(this));
-                }.bind(this)
-            );
         },
 
         /**
@@ -220,7 +174,8 @@ define(function () {
          */
         startLoadingOverlay: function() {
             var $container = this.sandbox.dom.createElement('<div class="'+ constants.loadingClass +'"/>'),
-                $loader = this.sandbox.dom.createElement('<div class="'+ constants.loaderClass +'" />')
+                $loader = this.sandbox.dom.createElement('<div class="'+ constants.loaderClass +'" />');
+            
             this.sandbox.dom.append(this.$el, $container);
             this.sandbox.once('husky.overlay.media-edit.loading.opened', function () {
                 this.sandbox.start([
@@ -270,10 +225,6 @@ define(function () {
                             {title: this.sandbox.translate(this.options.infoKey), data: this.$info},
                             {title: this.sandbox.translate(this.options.versionsKey), data: this.$versions}
                         ],
-                        languageChanger: {
-                            locales: ['en', 'de'],
-                            preSelected: this.media.locale
-                        },
                         openOnStart: true,
                         instanceName: 'media-edit',
                         propagateEvents: false,
@@ -328,10 +279,6 @@ define(function () {
                         el: $container,
                         title: this.sandbox.translate(this.options.multipleEditTitle),
                         data: this.$multiple,
-                        languageChanger: {
-                            locales: ['en', 'de'],
-                            preSelected: this.medias[0].locale
-                        },
                         openOnStart: true,
                         draggable: false,
                         propagateEvents: false,